backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.0.1
pydantic-settings==2.10.1
supabase==2.16.0
openai==1.99.9
python-jose[cryptography]==3.5.0
passlib==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
openpyxl==3.1.5
